Remove pushing from update homebrew script.
The script will soon be automatically invoked from a builder that
doesn't have ssh credentials for github but instead relies on gitcookies
to authenticate with git-on-borg. The push stepp will not work as it
currently is and is instead removed.

The script is now idempotent so the upcoming homebrew builder can be
rerun if a failed release is retried. It will instead simply not commit
any updates.

The Dart release script will be changed simultaneously when this change
lands to make sure the release procedure is uninterrupted.

Fix null safety issues in the parameter parsing and fix the runGit
function not handling failed invocations properly.

Bug: b/227748011