Automatically generate versioned formulas.
The stable releases now have versioned formulas automatically generated
and with a major.minor alias for that release series, so people can
continue to receive updates on their release or pin a very specific
build if it's needed.

The versioned formulas are currently up to date and this changelist
includes the result of running the script on the latest releases.

The --no-commit option is added for testing since it is very convenient
to just have the files not change but not having code changes mixed with
experimental outputs.

Delete the --revision option as it's not useful anymore and difficult
to implement with useful semantics.

Fixes: b/296366712
Change-Id: I8e04aba462f433f1dad513d482f090e0cee060dd
Reviewed-on: https://dart-review.googlesource.com/c/homebrew-dart/+/345120
Reviewed-by: William Hesse <whesse@google.com>
Commit-Queue: Jonas Termansen <sortie@google.com>